requirements.txt
requests
beautifulsoup4
python-dotenv
tqdm
lxml
html5lib